fixed a bug with the previous observer pattern implementation in which starting the VFCS would block, rather than start on a new thread.
