+++ src/main/java/com/york/asciiArtMaker/model/adapters/ImageFileAdapter.java
@@ -1,14 +1,23 @@
-package com.york.asciiArtMaker.adapters;
+package com.york.asciiArtMaker.model.adapters;
 
+import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
 import java.util.Optional;
 
-public class BufferedImageAdapter implements ImageSource {
+public class ImageFileAdapter implements ImageSource {
 
     private final BufferedImage bufferedImage;
+    private final String name;
 
-    public BufferedImageAdapter(BufferedImage bufferedImage) {
-        this.bufferedImage = bufferedImage;
+    public ImageFileAdapter(File file) {
+        try {
+            this.bufferedImage = ImageIO.read(file);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        this.name = file.getName();
     }
 
     @Override
@@ -23,7 +32,7 @@ public class BufferedImageAdapter implements ImageSource {
 
     @Override
     public Optional<String> getName() {
-        return Optional.empty();
+        return Optional.of(name);
     }
 
     @Override
@@ -36,4 +45,5 @@ public class BufferedImageAdapter implements ImageSource {
 
         return ImageSource.desaturate(a, r, g, b);
     }
+
 }
