+++ src/main/java/com/york/asciiArtMaker/model/adapters/MatAdapter.java
@@ -1,4 +1,4 @@
-package com.york.asciiArtMaker.adapters;
+package com.york.asciiArtMaker.model.adapters;
 
 import org.opencv.core.Mat;
