+++ src/main/java/com/york/asciiArtMaker/model/models/AppModel.java
@@ -1,7 +1,7 @@
-package com.york.asciiArtMaker.models;
+package com.york.asciiArtMaker.model.models;
 
-import com.york.asciiArtMaker.asciiArt.AsciiArtBuilder;
-import com.york.asciiArtMaker.asciiArt.AsciiImage;
+import com.york.asciiArtMaker.model.asciiArt.AsciiArtBuilder;
+import com.york.asciiArtMaker.model.asciiArt.AsciiImage;
 import com.york.asciiArtMaker.controller.MainController;
 
 import java.util.Optional;
